src/requirements.txt
pandas
python-dotenv
requests==2.31.0
requests-pkcs12==1.24
streamlit